Improve parsing of script arguments
Original version required -D, -I and -O options to be listed in exactly this order. While doable, this is somewhat counter-intuitive with typical Unix option processing expectations (e.g. `ls -l -a` is expected to behave the same as `ls -a -l`). This change introduces rudimentary options processing that allows specifying them in an arbitrary order (as long as they are before any consecutive command arguments).